requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart